pyproject.toml
[tool.poetry]
name = "gifmaker"
version = "0.1.0"
description = ""
authors = ["Elena Ivaschenko <elenaiv100@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pytube = "^12.1.0"
ffmpeg-python = "^0.2.0"
moviepy = "^1.0.3"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
